User ID Authentication, request and assignment of user roles and content editing functionality

+++ MunicipalPlatform-Java/MunicipalPlatform/MunicipalPlatform/src/main/java/it/unicam/cs/idsflsm/municipalplatform/application/services/contest/ContestService.java
@@ -4,13 +4,19 @@ import it.unicam.cs.idsflsm.municipalplatform.application.criterias.contest.Cont
 import it.unicam.cs.idsflsm.municipalplatform.application.mappers.contest.ContestMapper;
 import it.unicam.cs.idsflsm.municipalplatform.application.mappers.contest.ContributionMapper;
 import it.unicam.cs.idsflsm.municipalplatform.application.mappers.user.authenticated.AuthenticatedTouristMapper;
+import it.unicam.cs.idsflsm.municipalplatform.application.mappers.user.authenticated.GenericAuthenticatedUserMapper;
 import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.contest.ContestDto;
 import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.contest.ContributionDto;
 import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.user.authenticated.AuthenticatedTouristDto;
+import it.unicam.cs.idsflsm.municipalplatform.application.models.dtos.user.authenticated.AuthenticatedUserDto;
+import it.unicam.cs.idsflsm.municipalplatform.domain.entities.content.itinerary.Itinerary;
+import it.unicam.cs.idsflsm.municipalplatform.domain.entities.content.poi.POI;
 import it.unicam.cs.idsflsm.municipalplatform.domain.entities.contest.Contest;
 import it.unicam.cs.idsflsm.municipalplatform.domain.entities.contest.Contribution;
 import it.unicam.cs.idsflsm.municipalplatform.domain.entities.user.authenticated.AuthenticatedTourist;
+import it.unicam.cs.idsflsm.municipalplatform.domain.entities.user.authenticated.AuthenticatedUser;
 import it.unicam.cs.idsflsm.municipalplatform.domain.utilities.ContentState;
+import it.unicam.cs.idsflsm.municipalplatform.domain.utilities.ContestResult;
 import it.unicam.cs.idsflsm.municipalplatform.infrastructure.repositories.contest.IContestRepository;
 import it.unicam.cs.idsflsm.municipalplatform.infrastructure.repositories.contest.IContributionRepository;
 import jakarta.transaction.Transactional;
@@ -37,6 +43,10 @@ public class ContestService implements IContestService {
     public void saveInRepository(Contribution contribution) {
         _contributionRepository.save(contribution);
     }
+//    @Override
+//    public void deleteFromRepository(Contribution contribution) {
+//        _contributionRepository.delete(contribution);
+//    }
     @Override
     public List<ContestDto> getAllContests(Optional<Predicate<Contest>> predicate) {
         List<Contest> result = predicate.map(contestPredicate -> _contestRepository.findAll()
@@ -45,7 +55,7 @@ public class ContestService implements IContestService {
                     .collect(Collectors.toList()))
                 .orElseGet(_contestRepository::findAll);
         if (!result.isEmpty()) {
-            return ContestMapper.toDto(result);
+            return ContestMapper.toDto(result, true);
         } else {
             return null;
         }
@@ -53,17 +63,16 @@ public class ContestService implements IContestService {
     @Override
     public ContestDto getContestById(UUID id) {
         Optional<Contest> contest = _contestRepository.findById(id);
-        return contest.map(ContestMapper::toDto)
-                .orElse(null);
+        return contest.map(c -> ContestMapper.toDto(c, true)).orElse(null);
     }
     @Override
-    public boolean addContest(ContestDto contestDto) {
+    public ContestDto addContest(ContestDto contestDto) {
         if (!_contestRepository.existsById(contestDto.getId())) {
-            Contest contest = ContestMapper.toEntity(contestDto);
+            Contest contest = ContestMapper.toEntity(contestDto, true);
             _contestRepository.save(contest);
-            return true;
+            return contestDto;
         } else {
-            return false;
+            return null;
         }
     }
     @Override
@@ -94,7 +103,7 @@ public class ContestService implements IContestService {
                     .collect(Collectors.toList()))
                 .orElseGet(_contributionRepository::findAll);
         if (!result.isEmpty()) {
-            return ContributionMapper.toDto(result);
+            return ContributionMapper.toDto(result, true);
         } else {
             return null;
         }
@@ -102,22 +111,23 @@ public class ContestService implements IContestService {
     @Override
     public ContributionDto getContributionById(UUID id) {
         Optional<Contribution> contribution = _contributionRepository.findById(id);
-        return contribution.map(ContributionMapper::toDto)
-                .orElse(null);
+        return contribution.map(contribution1 -> ContributionMapper.toDto(contribution1, true)).orElse(null);
     }
     @Override
-    public boolean addContribution(UUID idContest, ContributionDto contributionDto) {
+    public ContributionDto addContribution(UUID idContest, UUID idParticipant, ContributionDto contributionDto) {
         Contest contest = _contestRepository.findById(idContest).orElse(null);
         if (contest != null) {
-            Contribution contribution = ContributionMapper.toEntity(contributionDto);
-            contribution.setState(ContentState.VALIDABLE);
-            contribution.setContest(contest);
-            contest.getContributions().add(contribution);
-            _contestRepository.save(contest);
-            return true;
-        } else {
-            return false;
+            var participants = contest.getParticipatingUsers();
+            if (participants.stream().anyMatch(authenticatedUser -> authenticatedUser.getId().equals(idParticipant))) {
+                Contribution contribution = ContributionMapper.toEntity(contributionDto, true);
+                contribution.setState(ContentState.VALIDABLE);
+                contribution.setContest(contest);
+                contest.getContributions().add(contribution);
+                _contestRepository.save(contest);
+                return ContributionMapper.toDto(contribution, true);
+            }
         }
+        return null;
     }
     @Override
     public boolean deleteContributionById(UUID id) {
@@ -144,43 +154,57 @@ public class ContestService implements IContestService {
 //        }
 //    }
     @Override
-    public boolean validateContribution(ContributionDto contributionDto, Optional<Predicate<Contribution>> predicate, boolean validate) {
-        Contribution contribution = ContributionMapper.toEntity(contributionDto);
-        assert contribution != null;
-        if (getAllContributions(predicate).get(0) != null) {
+    public ContributionDto validateContribution(UUID idContribution, boolean validate) {
+        Contribution contribution = _contributionRepository.findById(idContribution).orElse(null);
+        if (contribution != null && contribution.getState().equals(ContentState.VALIDABLE)) {
             ContentState newState = (validate) ? ContentState.UPLOADABLE : ContentState.REMOVABLE;
             contribution.setState(newState);
+            contribution.getContent().setState(newState);
             _contributionRepository.save(contribution);
-            return true;
-        } else {
-            return false;
-        }
-    }
-    @Override
-    public ContributionDto getWinnerContribution(UUID idContest) {
-        Optional<Contest> contest = _contestRepository.findById(idContest);
-        if (contest.isPresent()) {
-            Contribution result = contest.get().getContributions()
-                    .stream()
-                    .filter(ContributionCriteria.isWinner())
-                    .toList()
-                    .get(0);
-            return ContributionMapper.toDto(result);
+            return ContributionMapper.toDto(contribution, true);
         } else {
             return null;
         }
     }
     @Override
-    public Optional<Contribution> uploadContribution(ContributionDto contributionDto) {
-        Contribution contribution = ContributionMapper.toEntity(contributionDto);
-        if (contribution != null) {
-            // TODO : EXTRACTION OF POI / ITINERARY IN CONTRIBUTION, IN ORDER TO SAVE IT IN POI / ITINERARY REPO
-            _contributionRepository.delete(contribution);
-//            contribution.setState(ContentState.UPLOADED);
-//            _contributionRepository.save(contribution);
-            return Optional.of(contribution);
+    public ContributionDto getWinnerContribution(UUID idContest) {
+        Contest contest = _contestRepository.findById(idContest).orElse(null);
+        if (contest != null) {
+            Contribution result = contest.getContributions()
+                    .stream()
+                    .filter(ContributionCriteria.isWinner())
+                    .toList()
+                    .get(0);
+            return ContributionMapper.toDto(result, true);
         } else {
-            return Optional.empty();
+            return null;
+        }
+    }
+    @Override
+    public ContributionDto setWinnerContribution(UUID idContribution) {
+        Contribution contribution = _contributionRepository.findById(idContribution).orElse(null);
+        if (contribution != null && contribution.getState().equals(ContentState.UPLOADABLE) && contribution.getResult().equals(ContestResult.LOSER)) {
+            contribution.setResult(ContestResult.WINNER);
+            contribution.getContest().setHasWinner(true);
+            _contributionRepository.save(contribution);
+            return ContributionMapper.toDto(contribution, true);
+        } else {
+            return null;
+        }
+    }
+    @Override
+    public ContributionDto uploadContribution(UUID idContribution) {
+        Contribution contribution = _contributionRepository.findById(idContribution).orElse(null);
+        ContributionDto result = ContributionMapper.toDto(contribution, true);
+        if (contribution != null && contribution.getResult().equals(ContestResult.WINNER)) {
+            contribution.setNotNullPoi(null);
+            contribution.setNotNullItinerary(null);
+            contribution.getContest().getContributions().remove(contribution);
+            contribution.setContest(null);
+            _contributionRepository.delete(contribution);
+            return result;
+        } else {
+            return null;
         }
     }
     @Override
@@ -195,8 +219,10 @@ public class ContestService implements IContestService {
                 .toList();
         if (!result.isEmpty()) {
             result.forEach(contest -> contest.getContributions().forEach(contribution -> {
-                contribution.setState(ContentState.REMOVABLE);
-                _contributionRepository.save(contribution);
+                if (contribution.getResult().equals(ContestResult.LOSER)) {
+                    contribution.setState(ContentState.REMOVABLE);
+                    _contributionRepository.save(contribution);
+                }
             }));
             _contestRepository.saveAll(result);
             return true;
@@ -205,14 +231,14 @@ public class ContestService implements IContestService {
         }
     }
     @Override
-    public List<AuthenticatedTouristDto> getAllParticipants(UUID idContest, Optional<Predicate<AuthenticatedTourist>> predicate) {
+    public List<AuthenticatedUserDto> getAllParticipants(UUID idContest, Optional<Predicate<AuthenticatedUser>> predicate) {
         Contest contest = _contestRepository.findById(idContest).orElse(null);
         if (contest != null) {
-            List<AuthenticatedTourist> result = predicate.map(authenticatedTouristPredicate ->
-                contest.getParticipatingTourists().stream().filter(authenticatedTouristPredicate).toList()
-            ).orElseGet(contest::getParticipatingTourists);
+            List<AuthenticatedUser> result = predicate.map(authenticatedTouristPredicate ->
+                contest.getParticipatingUsers().stream().filter(authenticatedTouristPredicate).toList()
+            ).orElseGet(contest::getParticipatingUsers);
             if (!result.isEmpty()) {
-                return AuthenticatedTouristMapper.toDto(result);
+                return GenericAuthenticatedUserMapper.toDto(result, true);
             } else {
                 return null;
             }
@@ -221,43 +247,43 @@ public class ContestService implements IContestService {
         }
     }
     @Override
-    public AuthenticatedTouristDto getParticipantById(UUID idContest, UUID idAuthenticatedTourist) {
+    public AuthenticatedUserDto getParticipantById(UUID idContest, UUID idAuthenticatedUser) {
         Contest contest = _contestRepository.findById(idContest).orElse(null);
         if (contest != null) {
-            AuthenticatedTourist authenticatedTourist = contest.getParticipatingTourists()
+            AuthenticatedUser authenticatedUser = contest.getParticipatingUsers()
                     .stream()
-                    .filter(p -> p.getId().equals(idAuthenticatedTourist))
+                    .filter(p -> p.getId().equals(idAuthenticatedUser))
                     .findFirst()
                     .orElse(null);
-            return AuthenticatedTouristMapper.toDto(authenticatedTourist);
+            return GenericAuthenticatedUserMapper.toDto(authenticatedUser, true);
         } else {
             return null;
         }
     }
     @Override
-    public boolean addParticipant(UUID idContest, AuthenticatedTouristDto touristDto) {
-        AuthenticatedTourist authenticatedTourist = AuthenticatedTouristMapper.toEntity(touristDto);
-        assert authenticatedTourist != null;
-        Optional<Contest> contest = _contestRepository.findById(idContest);
-        if (contest.isPresent()) {
-            contest.get().getParticipatingTourists().add(authenticatedTourist);
-            _contestRepository.save(contest.get());
-            return true;
+    public AuthenticatedUserDto addParticipant(UUID idContest, AuthenticatedUserDto userDto) {
+        AuthenticatedUser authenticatedUser = GenericAuthenticatedUserMapper.toEntity(userDto, true);
+        Contest contest = _contestRepository.findById(idContest).orElse(null);
+        if (contest != null && !contest.getParticipatingUsers().contains(authenticatedUser)) {
+            contest.getParticipatingUsers().add(authenticatedUser);
+            authenticatedUser.getParticipatedContests().add(contest);
+            _contestRepository.save(contest);
+            return GenericAuthenticatedUserMapper.toDto(authenticatedUser, true);
         } else {
-            return false;
+            return null;
         }
     }
     @Override
-    public boolean deleteParticipant(UUID idContest, AuthenticatedTouristDto touristDto) {
-        AuthenticatedTourist authenticatedTourist = AuthenticatedTouristMapper.toEntity(touristDto);
-        assert authenticatedTourist != null;
-        Optional<Contest> contest = _contestRepository.findById(idContest);
-        if (contest.isPresent()) {
-            contest.get().getParticipatingTourists().remove(authenticatedTourist);
-            _contestRepository.save(contest.get());
-            return true;
+    public AuthenticatedUserDto deleteParticipant(UUID idContest, AuthenticatedUserDto userDto) {
+        AuthenticatedUser authenticatedUser = GenericAuthenticatedUserMapper.toEntity(userDto, true);
+        Contest contest = _contestRepository.findById(idContest).orElse(null);
+        if (contest != null && contest.getParticipatingUsers().contains(authenticatedUser)) {
+            contest.getParticipatingUsers().remove(authenticatedUser);
+            authenticatedUser.getParticipatedContests().remove(contest);
+            _contestRepository.save(contest);
+            return GenericAuthenticatedUserMapper.toDto(authenticatedUser, true);
         } else {
-            return false;
+            return null;
         }
     }
 }
